Allow a custom `animationCache` to be passed to `Sprite` constructor
The current default `animationCache` requies data to be stored and
maintained in the NAS2D library. By permitting an `animationCache`
parameter to the `Sprite` constructor, this cache can be moved external
to the library, where clients can be in control of the cache lifetime.

Given time for downstream adjustments, we could potentially remove the
now renamed `defaultAnimationCache` from the NAS2D library.

The now renamed `defaultAnimationCache` was pointed out by the Clang
warning flag `-Wglobal-constructors`.