Remove unpacked `connect`/`disconnect` overloads from `SignalSource`
A `DelegateX` object can be preconstructed at the point of call by
enclosing the two parameters with braces `{}`. That perhaps adds clarity
as well, since the `this` pointer and the member function pointer are
closely related, so it makes sense to pass them as a compound object.

Reducing the number of overloads makes it easier to add debug code to
just a single overload, which should help with the occasional event
handling issues.