Fix `Signal` memory corruption issue
Calling `Signal::connect` from within a signal callback called by
`Signal::emit` would update the underlying `std::vector`, which would
invalidate iterators. This was causing issues when control returned to
`Signal::emit` and it tried to call subsequent callbacks, which would be
done using the invalidated iterator on `std::vector` memory that had
already been freed. Since that memory was already freed, it was subject
to re-allocation and corruption.