Re-add `Timer::delta` method
This method is effectively a combination of `elapsedTicks` and
`adjustStartTick`. It is perhaps rather convenient in some contexts to
be able to easily call the combination of these two methods.

This method is not used within NAS2D, nor in OPHD, but is used by the
NAS2D test project.