Set Docker `workdir` at startup
The GitHub Actions docs recommend not using `WORKDIR` in the Dockerfile.
The working directory is set at startup, and is a mounted folder, so
will hide anything that was previously at that folder.

Local builds also mount the project folder into the image. Seems logical
to do similar as GitHub Actions and also set the `WORKDIR` setting at
startup to match.