Use `clang-format` to update preprocessor directive indentation
Technically the pre-processor directives are a separate language from
C++. The pre-processor (`cpp`) is run first, with the output then being
fed to the C++ compiler (`g++`). As such, it is often given a separate
indentation structure from that of C++ code. Historically, it was even
required for the `#` of a pre-processor directive to be in the first
column of a line.