Fix bug in `find` expression
The `find` expression was supposed to return all `.cpp` and `.h` files,
but to excludes files from the `Xml` subfolder.

The previous expression was returning only `.h` files that were not in
the `Xml` subfolder. The problem was a default logical "And" operation
was being performed between the `-prune` and the `-name '*.cpp'` check,
which had higher precedence than the `-o` "Or" operation with the `-name
'*.h'`. We needed the filename "Or" to have higher precedence, so they
would group together.

Additionally, with `-path` and `-prune`, the folder whose contents is
not traversed is returned itself as the only match. With "And" and a
filename filter, this means no results are returned. If instead you use
a logical "Or", then the excluded folder name is returned, since the
filename filter doesn't apply to it. This means `NAS2D/Xml` could be
passed to `clang-format`, which would respond with the warning message
"Is a directory". Of course, the previous `find` expression would filter
out that directory name because it didn't match `*.cpp` and there was an
implicit logical "And" between the two.