Add terminating commas to data lists and move initial brace
It seems `clang-format` will indent the whole block an extra level if
the initial brace is on a new line. Basically line continuations are
indented by an extra level.

Moving the brace to the end of the assignment line makes the data
definitions more consistent with existing data definitions used
elsewhere in the project.