Empty commit before `std::filesystem::path` changes
This commit can be used to test timings. It can be used as a reference
point to check if the `std::filesystem::path` changes result in longer
compile times. Part of that is because the `<filesystem>` include will
need to move from the .cpp file to the .h file, and so will cause
transitive header includes. Based on `-ftime-report`, we spend nearly
80% of compile time on source code parsing, which is strongly impacted
by transitive header includes.