Match alignment of `drawBox` and `drawBoxFilled`
Adjust alignment to center on pixels. OpenGL has a half pixel offset to represent the center of the pixel. Technically, we could subtract {0.5, 0.5} from the end point, though this makes the origin point disappear in the degenerate case of a 1x1 `Rectangle`. Not adjusting makes everything just work. Effectively, we exclude the lower right edges of the the `Rectangle` from being inside the box. This makes is very easy to align `Rectangle` objects so they meat up without any gap or overlap.