Update Windows platform detect to use `_WIN32`
The underscore prefixed name is in the implementation reserved namespace, which means the compiler can define it itself. This is common to both MSVC and MingW. The `WIN32` name is in a user namespace, and is defined within `<windows.h>` for MSVC. The MingW compiler can also define it, but only when not passing the `-std=c++??` option, which we do pass.