Convert error/warning message to exception
The only error we should really swallow in this function, is if the config file doesn't exist. In particular, it may be the first time the app is run, and so hasn't created the config file yet. Any other errors attempting to access the config file should be reported as errors.