Remove warning for config file not existing
The API design is simply that a missing config file is treated the same as an empty config file. No warning is necessary. In particular, it is quite normal for the config file to not exist on the first run of the app. In that case, default options are used.