Remove project setting for warnings as errors
We want warnings as errors for CI builds, which is enabled on CI with the `/warnAsError` option. We don't need to enforce this locally. If someone is testing a feature and just wants to know if something might work, they shouldn't have to fight with solving or silencing all warnings on a first go. When code gets to a PR stage, then it will matter, as the CI builds will fail if there are warnings.