Re-order Package.bat parameters
The `TargetPath` can potentially be determined from the other two, so could be made optional.

The `Platform` could potentially default to the current platform, though may be needed to support cross compiling. The `Configuration` could potentially also be defaulted to `Release` if this is used for release packaging, though a `Debug` package may also be built.