Use `vcpkg` in Manifest mode
Add a `vcpkg` Manifest file. Switch build to use the manifest file. This is a much more automated way for packages to be installed. There's also less wasted time once packages are installed, and less verbose console output when no updates are needed.

Remove old way of installing dependencies. We don't want the old `.bat` file left around to stagnate and confuse people.

The `--recurse` flag is not needed in Manifest mode. It is used in classic mode to uninstall packages that were installed with a different feature set. That's only necessary in a global context when packages may be in use by other projects, and so package removal should require some sort of confirmation.