Fix test-graphics project
It seems we had some holdover code from when we needed to pass `argv[0]` (the program name) to the `Filesystem` object. That parameter was removed, as it's no longer needed with the new implementation.

It seemed we forgot to update this project, and positional arguments meant the config file path was being set by the `argv[0]` value. This was causing errors during startup on Windows.