Update debug library names
Older version of `vcpkg` installed debug versions of the Google Test libraries with a "d" suffix. Now it seems there's no "d" suffix, and different configurations of the library are installed in separate folders.

Mind you, the debug builds linked with the current version of the library crash. It may be that the current built library is actually a release build, which may be the reason for dropping the "d" suffix.