Remove explicit step to install `vcpkg` dependencies
This will be handled automatically as part of the build.

The pre-install step was sort of nice in that it gave independent timing for package installation, versus building the project proper. However, there was a significant startup delay running the step, which was effectively duplicated once the build started. Removing the explicit pre-install step should remove the duplicated startup time for `vcpkg` to check if all dependent packages are up-to-date.