Remove incorrect include folder flag for test project
Includes specify the "NAS2D/" folder, so we should not add that to the include path. Doing so would make the header files available at two distinct paths. Instead, we rely on "./" being part of the include path, which allows "NAS2D/" to resolve correctly.

Looks like we forgot to remove this old flag when initially setting up the unit test project, which is not surprising, since the flag was placed in the compile command, rather than the pre-processor flags variable, where it should have been.