Specify NAS2D lib file directly when linking test projects
This makes parametes passed to the compiler/linker consistent with the dependency list, which can allow simplification using automatic variables.

It's also potentially simpler and faster to do it this way. No need to split the path into the folder and filename, and pass them as separate parameters. Plus, but not adding to the search path, there is no need to check an extra folder when looking for the other libraries.