Add comment about the importance of order of flags
Normally the `-U` (undefine) would be grouped with the pre-processor flags, however, the SDL2 flags are stored in `CXXFLAGS`, and they contain a `-D` (define), which is what needs to be undone, and so must come after that. Basically, `sdl2-config` outputs both pre-processor flags and compiler flags together, so things are a little less clean there.