Build test-graphic project with object file intermediates
Before we compiled all source files together, directly to an executable file. This is unlike other projects, which first compiled source files to object files, and then linked the object files together into the final executable.

Compiling everything together directly to an executable doesn't scale well with project size. In particular, there is no `ccache` caching of object outputs when that step is skipped. This meant any re-build of the test project, even when there were no changes, was a bit slow. Still generally within a second, but long enough of a delay to be noticeable.