Remove check for end of stream
There's no clear value in checking the `eof` bit here. The previous loop is designed to read up to the end of the string, and will never append an empty string when it ends with a delimiter. We always need to check for that and potentially add an empty string at the end.

In particular, if stream reading were to fail for a reason other than `eof` (which might not be possible with `istringstream`), it would still be correct, or at least not wrong, to check for an ending delimiter and append a blank string to the output.

It should be noted that the `istringstream` converts to a `bool` by checking `!fail()`. The `fail()` state is independent of the `eof` bit.

It should also be noted that `std::getline` functions as an `UnformattedInputFunction`. According to the documentation, that means if `eofbit` or `badbit` are set, then it will set `failbit` as well. Hence, `eof` leads to `fail`, which terminates the `getline` loop.

The check was first added in commit:
3cb767fb059b0dd7104a3a7091204211b8c8ce8f