Remove `shrink_to_fit()` call
There's actually very little value in attempting this here. Since we count delimiters up front, we basically already know how many output strings there will be. The only case where we overestimate is the empty string. It's hardly any memory in that case, as we only overestimate by 1. Plus, the `shrink_to_fit` function is not obligated to actually release any memory.

A better solution, might be to return early so as to never allocate that extra memory in the first place.