Adjust loop termination condition
This might be slightly easier to read, or maybe slightly easier for the compiler to optimize.

If the string ends with a delimiter, we treat it like there is an empty string following that delimiter. Hence, we need to loop one last time to add the empty string to the `result`.