Remove `RunCodeAnalysis` setting
Rather than force CodeAnalysis to run on every build, including builds of downstream projects, let's move this to on-demand. In particular, if CI builds run Code Analysis, and all builds need to pass before merging, we should expect the `main` branch to always be warning free.

Additionally, if we aren't forced to run Code Analysis on every build, we can experiment with running Code Analysis in parallel with a build.