Remove unused `Image` constructor
If we remove the `filePath` argument, this constructor overload would be ambiguous with the main constructor. That is, the `data` parameter of this constructor, and the `filePath` argument of the main constructor are both `std::string`, so would not produce unique overloads.

Technically this does remove some potentially useful functionality, is that we no longer have an overload that can process file data that comes from somewhere other than directly from the filesystem. Though there is potentially a workaround using one of the other data constructors, and possibly even with a derived class that has access to the protected static methods.