Rename dependencies from `.d` to `.dep`
This allows using the default dependency extension of `.d` when calling the compiler, and later renaming for use in the `makefile` to guard against corruption and timestamp issues. Using default values means they won't need to be specified explicitly.

This also makes the output a little easier to inspect using tab auto completion of filenames.