Add `workflow_dispatch` trigger to build workflow
This allows a run to be triggered on demand in the GitHub Actions web interface, without needing to push a change to a branch. This can be particularly useful when checking if changes to the build environment have broken the build, even though no code changes have been made in a long time.