Strip `query_` prefix from method names
There is sufficient context for function overloading to distinguish which method should be called. The main difference is if a `KeyModifier` is already known, or needs to be queried from the current keyboard state. When no `KeyModifier` is given, a reasonable assumption is to query for the information.

Removing the `_` makes the name more consistent with other methods, which prefer camelCase.