Remove no longer used Makefile for dependency source builds
This `Makefile` was relevant years ago when doing Ubuntu 16.04 builds, back when the stock `apt` packages for dependencies weren't up-to-date enough. It was not needed for Ubuntu 18.04 or newer, which shipped with newer versions of the needed `apt` packages.

----

PR where the file was first added and used:
https://github.com/lairworks/nas2d-core/pull/86

Merge commit:
1a5ef7294964e1292c317d4cc3e608082c40050f