Move Docker rules to separate `makefile`
By splitting these rules off from the main makefile, it makes it possible to watch for changes in to either the Dockerfiles or to the build rules in the `makefile` that generate Docker images from the Dockerfiles. This can be accomplished by watching for changes in the folder `docker/` to trigger builds of Docker images.

Having Docker images built using CI can potentially speed up the uplaod of large Docker images. Docker images can be quite large, particularly for the Mingw builds, which are close to 4GB. Having to upload them after a local build can take quite a bit of time.

It might make sense to use the GitHub Docker Registry rather than DockerHub, as that could allow for building and uploading all using GitHub infrastructure.