Split Docker image versions from `makefile`
These makes it easier to update separate images on separate branches, in parallel, and not encounter merge conflicts due to version number updates on adjacent lines.

This may also help in detecting which images need to be rebuilt, since we can watch for specific filenames, such as the Dockerfile or the associated version number file.