Add `branches-ignore` for `main` on `dockerBuild` workflow
The expected workflow is to make changes on a feature branch, where code will be tested and verified. Only once a new build image is created, and verified to run correctly, should the changes be merged into the `main` branch.

We don't want to trigger another build of the Docker image once the verified changes have been merged into the `main` branch. This would be counterproductive, as it forces a rebuild of something that should already exist, and may even introduce subtle changes in cases where installed packages have been updated between builds. That could result in using a different image than was tested and verified to work.