Make build dependent on detecting changes on the branch
We only run the workflow when there are changes to the docker folder (or the workflow itself), and we only want to run jobs for platforms that have an updated Docker image to build. Of course we can't detect that change at the job level, so instead we disable the steps to build and push the Docker image when there are no changes for the current platform.