Perform a `make clean` after scanning for warnings
Typically code changes would be tested, which may need a full compile with separate compiler flags, and often even with a different compiler. As a `make clean` was almost always run after `make show-warnings`, perhaps it makes sense to incorporate it as part of the command.

Speed wise, local Linux builds will likely already be using `ccache`, so there should be no real loss in rebuild speed from this clean. Rather, there is a gain in correctness, since we generally do need to rebuild.