Use `--output-sync` instead of `-j1` for faster warning scanning
The purpose of the `-j1` was to prevent interleaving of build output, which could confuse the scanning for warning messages. Unfortunately this was slow, particularly when full rebuilds were needed and no output was cached by `ccache`.

Instead of using `-j1`, the `--output-sync` flag is a better option. Output sync still allows for parallel builds, saving the output for each build target, and then printing each section of output together, without interleaving between build targets. This allows scanning the output for warning flags without getting confused by interleaved output, and without having to wait on a single core rebuild of the code.

Testing shows this saves quite a bit of time when generating warning flag output, particularly when a full rebuild is needed without `ccache` caching. Even when `ccache` already has output cached, there still seems to be a slight noticeable speedup.