Use `CURDIR` instead of `TopLevelFolder`
We probably want to run the Docker container from whatever folder the `make` command was run from, rather than always using the NAS2D folder. In particular, if a downstream project, such as OPHD, wants to compile it's code in the Docker container, it will be possible by running the `makefile` code from the downstream project folder.