Revert "Allow changes to the `makefile` to rebuild Docker images"
This reverts commit 120e9e2bd88f3788d53bc145504312f204aac28a.

Refactoring the `makefile` doesn't need to trigger a rebuild of all images. It was important during a name scheme change, since that was controlled by the `makefile`. Now that the name scheme change has been applied, we no longer need to trigger a build of all images for changes to the `makefile`.