Disable `cast-function-type` warning when compiling Glew
Normally there is a copious amount of warning messages when compiling Glew, which obscure any other messages as the output quickly fills the entire console scrollback buffer. This makes it harder to check on previous jobs for warnings, and harder to check on compile options and normal output messages.

Perhaps the original project could use some work to reduce the large number of warning messages. That's not part of this project though.