Add `.exe` suffix to executable Windows targets
Ideally the output filenames should match platform expectations. Normally the `.exe` suffix would be added anyway. Though the `makefile` rules reference the output filenames in a few places. Mostly the `.exe` suffix can be inferred. However, it is better for output names to match exactly, especially for `makefile` rules that need to determine if an output file is up-to-date.

Additionally, it was noted while looking into Wine upgrades that starting with Wine version 7, there seems to be trouble starting executable files when the `.exe` suffix is not explicitly specified. Failure to specify it can result in a hard to understand error message with code `c0000135` (`STATUS_DLL_NOT_FOUND` defined in `ntstatus.h`):
> wine: failed to open "../.build/Debug_Linux_test/test": c0000135

The same command works when the file to run is given an explicit suffix when passed to `wine`:
> "../.build/Debug_Linux_test/test.exe"