Add `user` so Wine has a writable prefix
Normally when run locally, it will run with the UID of the current user. For a single user Ubuntu system, that will likely be with UID `1000`. That user will need a writeable Wine prefix folder, which is typically `~/.wine`. If a matching user does not exist, it may try to write to `/.wine`, which would only be writeable to `root`. By creating a local user, we provide a writeable space for this common case, and can also cache the output of `wineboot`, which reduces Wine startup time when running a build in the container.