Don't cache the result of `wineboot`
Running `wineboot` creates an extra layer of about 400MB. Plus, on CircleCI the image runs as `root`, so we would need this cached for `root` rather than `user`. Additionally, switching to `user` and then back to `root` creates a Codacy warning. The warning is maybe a bit silly, but I suppose the `wineboot` layer caching is also a bit silly too.