Remove looped unit test
By removing the loop we also remove the `EXPECT_CALL` uses after a call has already been made to the mock. This is required to avoid undefined behavior.

The middle cases are less interesting, so we only preserve the boundary cases. Whether we use a time delta of 0, or one large enough to cycle through two stop frames, there should only ever be exactly one call to the handler. This is true even if the `advanceByTimeDelta` function is called a second time.