Allow specifying `sdl2-config` executable
There may be multiple `sdl2-config` executables present in a given environment. Having a variable for the executable file allows for easier customization of which executable to use.

A common case for this is doing cross compilation using Mingw to build a Windows executable from Linux. Normally the Linux environment would be setup for local building, and so `sdl2-config` would return flags suitable for a local Linux build. When doing cross compilation, there will be a separate binary install of SDL2 for Windows, which will have a separate `sdl2-config` executable, with Windows specific build flags and library paths.