Add a `CursorId` struct
This should help increase type safety. An `int` value could represent anything. A `CursorId` is specifically what it claims to be. There are no implicit conversions from `int` to `CursorId`, so this prevents the API from accepting arbitrary `int` values that may represent something else. It is however fairly simple to wrap an `int` value into a `CursorId`, making the conversion explicit.

Downstream projects, like OPHD have been using `enum` values for `cursorId`, and relying on the implicit conversion to `int`. It would be more type safe to use an `enum class`, with an explicit conversion to `CursorId` (rather than to `int`).