Add `Angle` struct
This wraps a simple radian measure.

This introduces a new type, distinct from `float`, which means we can increase type safety of functions taking angles, as an `Angle` parameter will not accept an arbitrary `float` value without an explicit conversion. Potentially this guards against parameters being passed in the wrong place, or selecting an inappropriate overload.