Expand `Angle` to full class with static constructors
In hindsight, using a `struct` with a single `public` member is not sufficient to ensure correct usage. Althugh it did introduce a new type, which could be used to increase type safety, it didn't address the distinction between degrees and radians. By using a `class`, and making the single field and constructor `private`, we can control construction using named `static` constructors. This ensures values are always explicitly specified as either degrees or radians.

Note that `Angle` is still a wrapper for a single primitive value. The aim of this class is to have a value that is easy and efficient to pass around, much like a primitive value, but with additional compile time type safety.