Add guard around call to `glDeleteTextures`
Despite documentation stating that `glDeleteTextures` can safely handle being passed `0` and will ignore them, that doesn't seem to be the case on MacOS. Passing in a `0` on MacOS leads to a segmentation fault.

In order to support a `null` `Font` object, we need to be able to pass `0` to `glDeleteTextures`.

Documentation:
https://registry.khronos.org/OpenGL-Refpages/es2.0/xhtml/glDeleteTextures.xml