Add static constructor `Font::null()`
A `null` `Font` object can be useful as a placeholder and for testing. Such a font would have nothing loaded, and could not print any text. It would however be usable as a `Font` object, and would be safe to call member functions on, without resulting in a crash or memory problem.

Here a named `static` constructor was chosen to make it obvious such an object was being created. By keeping the default construtor `protected`, we prevent accidentally creating `null` `Font` objects.