Split `EventHandler` enums to their own header files
Rarely are the full enum definitions needed.

They would be needed even less often if we translated events all the way to something like `onLeftClick`.

We include `EventHandler.h` from the enum headers, since they are normally nested in the `EventHandler` class. To enable that, a full definition of the `EventHandler` class needs to be seen.