Replace `Signal` with `Delegate` in `Fade`
This requires checking the `Delegate` is set before trying to call it. This is because `Delegate` defaults to an empty object with `nullptr` values, and setting the `Delegate` is optional. With `Signal`, the collection would have been empty, so it was just iterating over 0 elements. With the `Signal` code it was assumed that if a `Delegate` was set, it was a valid callable instance.