Update `makefile` to not use `sdl2-config`
The problem with `sdl2-config` is it would sometimes specify extra flags which were undesired, and there was no way to suppress the output of those flags. In some cases a defined macro value could be undefined, but there didn't seem to be a way to unlist a library. This sometimes caused problems with linking as not all projects need or want `SDL_main`.

It might be better to simply use `sdl2-config` in a manual way, and use it's output to guide how to use the `makefile` for a specific environment. At this point the `makefile` has sufficient customization points that any extra flags can be easily passed in. For example, `LDFLAGS_EXTRA` can be used to pass a library search path with the `-L` flag.