Add explicit `virtual` destructor to `Renderer`
We already inherit a `virtual` destructor from `Window`. However, by providing an explicit `override` here, we now have a `virtual` member with an out-of-line definition, meaning the virtual function table can be emitted in only the translation unit that defines it. This removes an instance of the Clang warning `-Wweak-vtables`.