Convert `Mix_Music` forward declare to include in `Music.h`
The `Mix_Music` typedef points to a `struct` that was renamed. There's no easy way to deal with that and maintain a forward declare. We would need to forward declare for one of two possible type names, before we've seen the include that would tell us which name it should be.

I suppose we could add some pointer casting, but that should be avoided.

Considering how infrequently the `Music.h` header is included by other files, we're probably fine doing a full SDL include for now. Once things have stabilized on the new version we could eventually switch back to a forward declare. As a bonus, that forward declare will then be for a name that is not reserved.