Add explicit `static_cast` to `OutputType`
This is particularly important when `OutputType` is a smaller type than `InputType`. In such a case, the intermediate calculations will be promoted to the type of higher rank, which would be the `InputType`. Thus a conversion to a smaller `OutputType` could lead to warnings about implicit conversions and possible loss of data.