Update `scaleLinear` to better handle wrap around cases
In particular, when there is a negative slope the previous code caused wrap around when `OutputType` was unsigned, and then performed multiplication and division on the result. Performing multiplication or division after a wrap around will produce unexpected results. The updated code combines two separately linear scaled values without relying on wrap around of the `OutputType`. Effectively it is a sum of shifting weighted averages of the two values with zero. By averaging them separately we avoid the wrap around.

A note on types:

When two integer values are operated on, they are promoted to at least `int` before the operation. The result is then the promoted type. For integer values, the following is at least an `int` type, or larger.
```cpp
using PromotedInputType = decltype(domainPoint1 - domainPoint2);
```

When an integer value and a floating point value are operated on, they are promoted to the floating point type. For mixed `InputType` and `OutputType` values, the following is of the floating point type.
```cpp
using PromotedOutputType = decltype((domainPoint1 - domainPoint2) * rangePoint1);
```

It should be noted that converting from integer types to floating point types is a narrowing conversion (unless the input value is `constexpr` and fits within the output type). Thus any extracted sub-expression would need to allow for narrowing conversions. For instance, to get a compatible promoted type to multiply or divide with the `OutputType`, we would want to use rounded parenthesis, rather than braces.
```cpp
const auto domainWidth = PromotedOutputType(domainPoint2 - domainPoint1);
```

Of course doing everything as one big calculation, without extracted intermediate sub-expressions avoids making these details explicit.