Use elapsed time to determine fade complete
Using the time values is perhaps easier to read and understand. Additionally, the change makes it easier to then use `scaleLinear` for the main calculation.

There is maybe a slight change in behavior here, regarding rounding errors in determining when a fade is complete. With the change we wait the full elapsed time before declaring the fade complete. Previously, once the output was rounded to max intensity, the fade was considered complete, even if it hadn't quite yet reached the full fade time. Arguably it was more efficient to stop early, once there was no more calculation to do, though if something else was waiting on the fade to complete, it could have triggered early if the calculation finished early. With the update, everything should finish at the specified time.