Replace `switch` with cacaded `if`/`else if`
Arguably the `switch` is better, since there is no repetition of the `key` value being checked. Though the code is longer, and more prone to errors due to missing `break` or missing block scopes `{}` in case local variables need to be introduced.

What tips the balance was the Clang warning `-Wswitch-enum` which this solves. When not all cases of an enum are covered it can trigger that Clang warning. We basically never want to cover every `KeyCode` value.