Extract default graphics and audio config to functions
We use functions since the values are not `constexpr` compatible, and we don't want to add global objects that require construction.

Potentially we might want to match original defaults for just one section with updates for another section. Such re-use is easier when the different sections can be referred to individually.