Make use of `PkgConfig` optional
If the `PkgConfig` specified tool can not be found, then simply skip it. We don't want to see warnings in those cases. Assume the environment is setup specifically for the type of build being done, and has minimal tooling installed. In particular, assume the environment has a specific `Toolchain` installed, and no tooling to support alternative toolchains. This is the case for the CI builds, which are run from Docker images, with only the one toolchain installed.