Add allowed and known set of Clang warnings
Add sets of disabled warning flags for allowed Clang warnings and known Clang warnings. These warnings may be of occasional interest, so we may want to enable to disable them separately from other flags. These flags are not part of a failure set for warnings, but may be of interest when scanning to show known warnings.

Allowed warnings are distinct from warnings that will never be of any interest, such as warning about code being incompatible with obsolete C++ standards which we don't target. Known warnings are things we might like to fix, but haven't addressed yet.